+++ packages/eslint-config/nextjs.js
@@ -57,6 +57,11 @@ const config = [
         plugins: {
             react: react,
         },
+        settings: {
+            react: {
+                version: "detect",
+            },
+        },
         rules: {
             "react/display-name": "off",
             "react/jsx-curly-brace-presence": "error",